TG-92: Fix folder structure and outdated implementations
