--- src/Components/Projects/CreateProject/createProject.module.css
@@ -1,94 +1,0 @@
-* {
-	box-sizing: border-box;
-}
-
-a {
-	color: black;
-	text-decoration: none;
-}
-
-.navContainer {
-	display: flex;
-	align-items: center;
-	gap: 12px;
-}
-
-.formContainer {
-	margin: 50px auto;
-	padding: 20px;
-	border-radius: 10px;
-	background-color: white;
-	min-width: 300px;
-	max-width: 60%;
-}
-
-.form {
-	display: flex;
-	flex-direction: column;
-	padding: 0;
-	width: 100%;
-}
-
-.form input {
-	margin: 4px 0;
-	padding: 4px;
-	width: 100%;
-}
-
-.buttonContainer {
-	display: flex;
-	gap: 15px;
-	padding: 0;
-	width: 100%;
-	justify-content: space-between;
-}
-
-.form button {
-	color: white;
-	padding: 8px;
-	width: 50%;
-	margin-top: 8px;
-}
-
-.form button:first-child {
-	background-color: red;
-	border: none;
-	border-radius: 5px;
-}
-
-.form button:last-child {
-	background-color: green;
-	border: none;
-	border-radius: 5px;
-}
-
-.selectedItem {
-	margin: 0;
-	padding: 10px;
-	width: 100%;
-	background-color: rgb(236, 255, 236);
-}
-
-.selectedItem:not(:last-child) {
-	box-shadow: 0px -1px 0px 0px #696 inset;
-}
-
-.notSelectedItem {
-	margin: 0;
-	padding: 10px;
-	width: 100%;
-	width: 100%;
-	background-color: #fafafa;
-}
-
-.notSelectedItem:not(:last-child) {
-	box-shadow: 0px -1px 0px 0px #666 inset;
-}
-
-.selectedItem:hover {
-	background-color: rgb(247, 199, 199);
-}
-
-.notSelectedItem:hover {
-	background-color: #eee;
-}
