--- src/Components/Projects/List/ListItem/ListItemDropdown/DropdownItem/index.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const DropdownItem = (props) => {
-  return <a href={props.toPage}>{props.message}</a>;
-};
-
-export default DropdownItem;
