--- src/Components/Projects/List/ListItem/ListItemDropdown/listItemDropdown.module.css
@@ -1,42 +1,0 @@
-* {
-	box-sizing: border-box;
-}
-
-.dropdown {
-	position: relative;
-}
-
-.dropdownContent {
-	display: none;
-	position: absolute;
-	width: 100px;
-	z-index: 10;
-}
-
-.dropdown:hover .dropdownContent {
-	display: block;
-}
-
-.dropdownContent > a {
-	color: black;
-	text-decoration: none;
-	line-height: 20px;
-	padding: 10px 4px;
-	display: block;
-	margin: 0;
-	height: 40px;
-	text-justify: center;
-	text-overflow: ellipsis;
-	overflow: hidden;
-	white-space: nowrap;
-}
-
-.dropdownContent :not(:last-child) {
-	background-color: #d0d4d9ee;
-	box-shadow: 0px -1px 0px 0px #c0c4c9ee inset;
-}
-
-.dropdownContent :last-child {
-	background-color: #d0d4d9ee;
-	border-radius: 0 0 5px 5px;
-}
