--- src/Components/Projects/List/ListItem/listItem.module.css
@@ -1,19 +1,0 @@
-.row {
-	height: 80px;
-}
-
-.row > td {
-	padding: 10px;
-}
-
-.edit, .delete {
-	transition: .05s;
-}
-
-.edit:hover {
-	color: blue;
-}
-
-.delete:hover {
-	color: red;
-}
