+++ src/Components/Projects/index.jsx
@@ -1,11 +1,13 @@
 import React from 'react';
 import List from './List';
-import CreateProject from './CreateProject';
 import styles from './projects.module.css';
 
-function Projects(props) {
-  const { form } = props;
-  return <section className={styles.container}>{form ? <CreateProject /> : <List />}</section>;
+function Projects() {
+  return (
+    <section className={styles.container}>
+      <List />
+    </section>
+  );
 }
 
 export default Projects;
