+++ src/Components/Projects/projects.module.css
@@ -1,4 +1,4 @@
-.container > * {
+.container {
   flex: 1;
   display: flex;
   flex-direction: column;
