+++ src/Components/Shared/Form/form.module.css
@@ -3,7 +3,7 @@
 	font-weight: bold;
 	color: #373867;
 	display: flex;
-	flex-direction: row;
+	flex-direction: column;
 	justify-content: center;
 }
 h2{
