+++ src/Components/Shared/Modal/modal.module.css
@@ -24,9 +24,9 @@
 	box-shadow: 0px 10px 23px rgba(0, 0, 0, 0.12);
 	padding: 16px;
 	margin: auto;
-	min-width: 360px;
+	min-width: 350px;
 	max-width: 800px;
-	width: fit-content;
+	max-height: 90vh;
 }
 
 .modalHeader {
@@ -48,5 +48,5 @@
 .modalBody {
  overflow: auto;
  height: max-content;
- max-height: calc(100vh - 48px);
+ max-height: calc(90vh - 75px);
 }
